Separate tui app logic from row parsing logic
